refactor the changeHandlers from BasicSettings.js component to external helpers file
